Renaming base test folder
