+++ tests/test_drivers.py
@@ -5,7 +5,7 @@ import sys  # fix import errors
 import os
 import json
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from tests.base_config import ConfigTestCase
+from tests.base_config_tests import ConfigTestCase
 
 class DriversEndpoint(ConfigTestCase):
     """This class represents driver test cases"""
